[chore] #52 파일 분리 작업
