+++ SniffMeet/SniffMeet/Source/Auth/CreateProfile/Interactor/SaveInfoUseCase.swift
@@ -6,26 +6,6 @@
 //
 import Foundation
 
-protocol DataStorable {
-    func storeData(data: Encodable) throws
-}
-
-// TODO: -  dataManager를 주입받을 수 있도록 수정 예정
-final class LocalDataManager: DataStorable {
-    private let dataManager = UserDefaultsManager(userDefaults: UserDefaults(suiteName: "demo")!,
-                                                  jsonEncoder: JSONEncoder(),
-                                                  jsonDecoder: JSONDecoder())
-    func storeData(data: any Encodable) throws {
-        try dataManager.set(value: data, forKey: UserDefaultKey.dogInfo)
-    }
-}
-
-extension LocalDataManager {
-    private enum UserDefaultKey {
-        static let dogInfo: String = "dogInfo"
-    }
-}
-
 protocol StoreDogInfoUseCase {
     func execute(dog: Dog) throws
 }
